Add Compiler for Twig Engine
See #72, adds a Compiler for Laravel. Not sure if `isExpired` and
`getCompiledPath` work as intended by the interface, but I don't think
they are ever called.
Also might think about where to put the Environment, I removed it now
from the engine, but maybe let the compiler also load the templates? Or
inject TwigEnvironment directly in the engine?